match java packages to sonatype canonical namespace
