+++ dependencyversion-maven-plugin/src/main/java/io/reformanda/semper/maven/dependencyversion/SetContains.java
@@ -1,0 +1,32 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package io.reformanda.semper.maven.dependencyversion;
+
+import java.util.Collections;
+import java.util.Optional;
+import java.util.Set;
+import java.util.function.Function;
+import java.util.function.Predicate;
+
+public class SetContains <T> implements Function<Set<T>, Predicate<T>> {
+
+    @Override
+    public Predicate<T> apply(Set<T> ts) {
+        final Set<T> resultSet = Optional.ofNullable(ts).orElse(Collections.emptySet());
+        return resultSet.isEmpty() ? s -> false : resultSet::contains;
+    }
+}
